pyproject.toml
[tool.poetry]
name = "case_management"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.2"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"
werkzeug = "^1.0.1"
whitenoise = "^5.2.0"
sentry-sdk = "^0.18.0"
django-extensions = "^3.0.9"
pytest-cov = "^2.10.1"
html5lib = "^1.1"
codecov = "^2.1.9"
gunicorn = "^20.0.4"
djangorestframework = "^3.12.1"
gevent = "^20.9.0"
django-cors-headers = "^3.7.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
